Add varicode decoding and headers for callbacks
Previously varicode decoding was skipped, so we didn't get the
data channel data that is used by freedv to decode callsigns. The code was
copied from freedv svn r1155. This is not sent to the Java side at this point.

Function prototypes were not provided for functions in the libandroidfreedv
files that were called between files. They now are.